corregit el makefile y les noms d eles funcions

+++ libft/ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: doller-m <doller-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/03 12:57:40 by doller-m          #+#    #+#             */
-/*   Updated: 2023/05/16 15:29:15 by doller-m         ###   ########.fr       */
+/*   Updated: 2023/05/16 16:56:34 by doller-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,9 +14,9 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	int				i;
-	unsigned char	*strsrc;
-	unsigned char	*strdst;
+	unsigned long int	i;
+	unsigned char		*strsrc;
+	unsigned char		*strdst;
 
 	i = 0;
 	strsrc = (unsigned char *)src;
