+++ libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: doller-m <doller-m@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/05 13:06:02 by doller-m          #+#    #+#             */
-/*   Updated: 2023/05/16 15:25:23 by doller-m         ###   ########.fr       */
+/*   Updated: 2023/05/16 17:00:32 by doller-m         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,10 +14,10 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int		i;
-	int		j;
-	char	*aguja;
-	char	*pajar;
+	int					i;
+	unsigned long int	j;
+	char				*aguja;
+	char				*pajar;
 
 	aguja = (char *)needle;
 	pajar = (char *)haystack;
